+++ jsconfig.json
@@ -1,0 +1,8 @@
+{
+    "compilerOptions": {
+      "moduleResolution": "node",
+      "module": "es6",
+      "target": "es6"
+    },
+    "include": ["src/**/*"]
+  }
