+++ src/features/MagicBox/components/Box/box.scss
@@ -1,0 +1,11 @@
+.box {
+    display: inline-block;
+    padding: 2rem;
+    margin: 1rem;
+
+    color: #fff;
+    font-size: 24px;
+    font-weight: bold;
+
+
+}
